requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.3
decorator==4.3.0
Django==2.1
djangorestframework==3.8.2
gevent==1.3.5
gevent-eventemitter==2.0
greenlet==0.4.14
httplib2==0.11.3
idna==2.7
kombu==4.2.1
mysqlclient==1.3.13
protobuf==3.6.0
pycparser==2.18
pytz==2018.5
redis==2.10.6
requests==2.19.1
six==1.11.0
steam==0.9.0
urllib3==1.23
validators==0.12.2
vdf==2.3
vine==1.1.4
xmltodict==0.11.0
